+++ .prettierrc.js
@@ -6,4 +6,3 @@
   "tabWidth": 2,
   "useTabs": false
 }
-a
