+++ index.js
@@ -3,7 +3,7 @@
  */
 
 import { AppRegistry } from 'react-native';
-import App from './App';
 import { name as appName } from './app.json';
+import App from "./src/App";
 
 AppRegistry.registerComponent(appName, () => App);
