+++ package.json
@@ -12,9 +12,12 @@
     "compile": "tsc --noEmit"
   },
   "dependencies": {
-    "@react-native/new-app-screen": "0.80.2",
+    "@react-navigation/native": "^7.1.17",
+    "@react-navigation/native-stack": "^7.3.26",
     "react": "19.1.0",
-    "react-native": "0.80.2"
+    "react-native": "0.80.2",
+    "react-native-safe-area-context": "^5.6.1",
+    "react-native-screens": "^4.16.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
