+++ src/navigation/ MainStack.tsx
@@ -1,0 +1,20 @@
+import * as React from 'react'
+import { NavigationContainer } from '@react-navigation/native'
+import { createNativeStackNavigator } from '@react-navigation/native-stack'
+import HomeScreen from "../screens/HomeScreen"
+
+
+
+const Stack = createNativeStackNavigator()
+
+const MainStack = () => {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator>
+      <Stack.Screen name="Home" component={HomeScreen} />
+    </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
+
+export default MainStack;
