+++ src/screens/HomeScreen.tsx
@@ -1,0 +1,11 @@
+import { Text, View } from "react-native"
+
+const  HomeScreen = () => {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Home Screen</Text>
+    </View>
+  );
+}
+
+export default HomeScreen
